% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micss.R
\name{p.val.kappa}
\alias{p.val.kappa}
\title{P-values}
\usage{
p.val.kappa(x, t, alpha)
}
\arguments{
\item{x}{Value of the statistic.}

\item{t}{Sample size.}

\item{alpha}{Value of the tail index between 2 and 4.}
}
\value{
p-value.
}
\description{
Computes p-values for the \link{kappa_test} statistic
}
\references{
J.L. Carrion-i-Silvestre & A. Sansó (2026): Testing for Constant Unconditional Variance in Heavy-Tailed Time Series. Communications in Statistics - Simulation and Computation. doi:10.1080/03610918.2026.2615207
}
\seealso{
\link{cv.kappa}
}
\keyword{internal}
