% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micss.R
\name{print.micss}
\alias{print.micss}
\title{print.micss}
\usage{
\method{print}{micss}(x, ...)
}
\arguments{
\item{x}{An object with the output of the \link{micss} algorithm.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
{No return value. It prints the output of \link{micss}}
}
\description{
Prints the output of \link{micss}.
}
\references{
J.L. Carrion-i-Silvestre & A. Sansó (2026): Testing for Constant Unconditional Variance in Heavy-Tailed Time Series. Communications in Statistics - Simulation and Computation. doi:10.1080/03610918.2026.2615207
}
\examples{
set.seed(2)
e <- c(stats::rnorm(200),3*stats::rnorm(200))
o <- micss(e)
print.micss(o)
}
