% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micss.R
\name{sr_loc}
\alias{sr_loc}
\title{sr_loc}
\usage{
sr_loc(t, alpha, lmax = NULL)
}
\arguments{
\item{t}{Sample size}

\item{alpha}{Value of the tail index between 2 and 4}

\item{lmax}{Maximum lag to be used for the long-run estimation of the fourth order moment of the
innovations. If not specified it is generated automatically depending on the sample size.}
}
\value{
Location parameter
}
\description{
Computes the location parameter for the GEV approximation
to the distribution of \link{kappa_test} statistic
}
\details{
used internally by \link{cv.kappa} and \link{p.val.kappa}
}
\references{
J.L. Carrion-i-Silvestre & A. Sansó (2026): Testing for Constant Unconditional Variance in Heavy-Tailed Time Series. Communications in Statistics - Simulation and Computation. doi:10.1080/03610918.2026.2615207
}
\keyword{internal}
