% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micss.R
\name{step2c}
\alias{step2c}
\title{step2c}
\usage{
step2c(e, bb, sig.lev, kmax, alpha)
}
\arguments{
\item{e}{Stationary variable on which the constancy of unconditional variance is tested.}

\item{bb}{The output of function \link{steps1.to.2b}.}

\item{sig.lev}{Significance level.}

\item{kmax}{Maximum lag to be used for the long-run estimation of the fourth order moment of the innovations.}

\item{alpha}{Tail index.}
}
\value{
\itemize{
  \item \code{nb}: Number of breaks.
  \item \code{tb}: Time of the breaks.
}
}
\description{
Computes step 2c of the \link{icss} Algorithm.
}
\details{
Used internally by \link{icss}.
}
\references{
C. Inclan & G.C. Tiao (1994): Use of Cumulative Sums of Squares for Retrospective Detection of Changes of Variance. Journal of the American Statistical Association 89, 913-923.
}
\keyword{internal}
