% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micss.R
\name{taula.icss}
\alias{taula.icss}
\title{taula.icss}
\usage{
taula.icss(x)
}
\arguments{
\item{x}{An object with the output of \link{icss}.}
}
\value{
Matrix with the different periods and estimated variances.
}
\description{
When there are breaks in the unconditional variance, it
creates a matrix with the different periods and estimated variances.
}
\details{
Used internally by \link{print.icss}. If there are no break returns NA value.
}
\keyword{internal}
