% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micss.R
\name{taula.micss}
\alias{taula.micss}
\title{taula.micss}
\usage{
taula.micss(x, alpha)
}
\arguments{
\item{x}{An object with the output of \link{icss}.}

\item{alpha}{Value of the tail index between 2 and 4}
}
\value{
Matrix with the different periods, estimated variances, values of
 \link{kappa_test} and \link{p.val.kappa}.
}
\description{
When there are breaks in the unconditional variance, it
creates a matrix with the different periods, estimated variances, values of
 \link{kappa_test} and \link{p.val.kappa}.
}
\details{
Used internally by \link{print.micss}.
}
\keyword{internal}
