% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micss.R
\name{var.est.icss}
\alias{var.est.icss}
\title{var.est.icss}
\usage{
var.est.icss(x)
}
\arguments{
\item{x}{An object with the output of \link{icss} or \link{micss}.}
}
\value{
A vector with the variances.
}
\description{
Computes the variance of each period according to the breaks found with \link{icss} or \link{micss}.
}
\keyword{internal}
