% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micss.R
\name{cv.kappa}
\alias{cv.kappa}
\title{Critical Values}
\usage{
cv.kappa(t, alpha, sig.lev)
}
\arguments{
\item{t}{Sample size.}

\item{alpha}{Value of the tail index between 2 and 4.}

\item{sig.lev}{Significance level.}
}
\value{
Critical value.
}
\description{
Computes critical values for the \link{kappa_test} statistic
}
\references{
J.L. Carrion-i-Silvestre & A. Sansó (2023): Generalized Extreme Value Approximation to the CUMSUMQ Test for Constant Unconditional Variance in Heavy-Tailed Time Series.
}
\seealso{
\link{p.val.kappa}
}
\keyword{internal}
