% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micss.R
\name{print.icss}
\alias{print.icss}
\title{print.icss}
\usage{
\method{print}{icss}(x, ...)
}
\arguments{
\item{x}{An object with the output of the \link{icss} algorithm.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
{No return value. It prints the output of \link{icss}}
}
\description{
Prints the output of \link{icss}.
}
\details{
Used internally by \link{icss}.
}
\examples{
set.seed(2)
e <- c(stats::rnorm(200),3*stats::rnorm(200))
o <- icss(e)
print.icss(o)
}
